+++ docs/.vuepress/theme/index.js
@@ -1,12 +1,14 @@
-import { resolve } from 'path';
+import { getDirname, path } from '@vuepress/utils';
 import { themeDataPlugin } from '@vuepress/plugin-theme-data';
 import { tocPlugin } from '@vuepress/plugin-toc';
 import { activeHeaderLinksPlugin } from '@vuepress/plugin-active-header-links';
 
+const __dirname = getDirname(import.meta.url);
+
 export const dialtoneVuepressTheme = (options) => {
   return {
     name: '@dialpad/vuepress-theme-dialtone',
-    clientConfigFile: resolve(__dirname, 'client.js'),
+    clientConfigFile: path.resolve(__dirname, 'client.js'),
     plugins: [
       themeDataPlugin({
         themeData: options,
@@ -25,3 +27,5 @@ export const dialtoneVuepressTheme = (options) => {
     ],
   };
 };
+
+export default dialtoneVuepressTheme;
