+++ package.json
@@ -74,7 +74,9 @@
     "access": "public"
   },
   "dependencies": {
-    "@dialpad/dialtone-icons": "0.0.7-vue3",
+    "@dialpad/dialtone-combinator": "^0.3.1",
+    "@dialpad/dialtone-icons": "0.0.9-vue3",
+    "@dialpad/dialtone-vue": "^3.26.0",
     "@docsearch/js": "^3.2.1",
     "date-fns": "^2.29.3",
     "docopt": "^0.6.2",
@@ -84,8 +86,6 @@
     "@commitlint/cli": "^17.1.2",
     "@commitlint/config-conventional": "^17.1.0",
     "@dialpad/conventional-changelog-angular": "^1.0.0",
-    "@dialpad/dialtone-combinator": "^0.3.1",
-    "@dialpad/dialtone-vue": "^3.25.1",
     "@dialpad/postcss-responsive-variations": "^1.1.3",
     "@semantic-release/changelog": "^6.0.1",
     "@semantic-release/git": "^10.0.1",
